requirements.txt
cn_clip==1.5.1
numpy==1.24.2
opencv_python==4.9.0.80
pandas==2.0.3
Pillow==9.4.0
Pillow==11.0.0
Requests==2.32.3
scikit_learn==1.3.2
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.61.2
transformers==4.36.2
